+++ OneBlink.SDK/models/FormSubmissionEvent.cs
@@ -104,11 +104,13 @@ namespace OneBlink.SDK.Model
             return civicaCrm;
         }
 
-        public static FormSubmissionEvent CreateSchedulingSubmissionEvent(string nylasAccountId, long nylasSchedulingPageId, bool isDraft = false)
+        public static FormSubmissionEvent CreateSchedulingSubmissionEvent(string nylasAccountId, long nylasSchedulingPageId, Guid nameElementId = default(Guid), Guid emailElementId = default(Guid),bool isDraft = false)
         {
             FormSubmissionEventConfigration fseconfig = new FormSubmissionEventConfigration();
             fseconfig.nylasAccountId = nylasAccountId;
             fseconfig.nylasSchedulingPageId = nylasSchedulingPageId;
+            fseconfig.nameElementId = nameElementId;
+            fseconfig.emailElementId = emailElementId;
             FormSubmissionEvent schedulingEvent = new FormSubmissionEvent();
             schedulingEvent.type = "SCHEDULING";
             schedulingEvent.configuration = fseconfig;